+++ pillarbox-demo-shared/src/main/java/ch/srgssr/pillarbox/demo/shared/ui/TalkBack.kt
@@ -36,3 +36,27 @@ fun rememberIsTouchExplorationEnabled(): Boolean {
 
     return isTouchExplorationEnabled
 }
+
+/**
+ * A composable function that returns a boolean indicating whether TalkBack is currently enabled.
+ *
+ * This function uses a [DisposableEffect] to register an [AccessibilityManager.AccessibilityStateChangeListener]
+ * that updates the state of the composable when the accessibility state changes.
+ *
+ * @return `true` if TalkBack is enabled, `false` otherwise.
+ */
+@Composable
+fun rememberIsTalkBackEnabled(): Boolean {
+    val accessibilityManager = LocalContext.current.getSystemService<AccessibilityManager>() ?: return false
+    val (isTalkBackEnabled, setTalkBackEnabled) = remember {
+        mutableStateOf(accessibilityManager.isEnabled)
+    }
+    DisposableEffect(Unit) {
+        val l = AccessibilityManager.AccessibilityStateChangeListener(setTalkBackEnabled)
+        accessibilityManager.addAccessibilityStateChangeListener(l)
+        onDispose {
+            accessibilityManager.removeAccessibilityStateChangeListener(l)
+        }
+    }
+    return isTalkBackEnabled
+}
